+++ test/tnep.jl
@@ -47,6 +47,27 @@ end
     end
 end
 
+
+@testset "test qc tnep" begin
+    @testset "3-bus case" begin
+        result = run_tnep("../test/data/matpower/case3_tnep.m", QCWRPowerModel, pajarito_solver)
+
+        check_tnep_status(result["solution"])
+
+        @test result["status"] == :Optimal
+        @test isapprox(result["objective"], 2; atol = 1e-2)
+    end
+    @testset "5-bus rts case" begin
+        result = run_tnep("../test/data/matpower/case5_tnep.m", QCWRPowerModel, pajarito_solver)
+
+        check_tnep_status(result["solution"])
+
+        @test result["status"] == :Optimal
+        @test isapprox(result["objective"], 1; atol = 1e-2)
+    end
+end
+
+
 @testset "test dc tnep" begin
     @testset "3-bus case" begin
         result = run_tnep("../test/data/matpower/case3_tnep.m", DCPPowerModel, pajarito_solver)