+++ CHANGELOG.md
@@ -5,9 +5,10 @@ PowerModels.jl Change Log
 - nothing
 
 ### v0.8.1
+- Strengthened the QCWRTri Power Flow formulation
 - Added support for implicit single conductor to buspairs data
 - Made add_setpoint more flexible when working with a mixture of data types
-- Improved the QC lambda-based relaxation
+- Fixed a bug in TNEP voltage variable definitions
 
 ### v0.8.0
 - Added support for asymmetric line charging in all formulations